// src/api/generated/model/categoryResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */

export interface CategoryResponse {
	id?: string;
	slug?: string;
	title?: string;
	image?: string;
	featured?: boolean;
}